+++ API_tests/src/test/java/stories/UserStory/HiredJobs.java
@@ -6,13 +6,11 @@ import org.testng.annotations.BeforeTest;
 import org.testng.annotations.Test;
 import org.testng.reporters.jq.INavigatorPanel;
 import stories.managers.SettingsManager;
+import stories.model.shademodel.core.model.accountmodel.UserModel;
 import stories.model.shademodel.core.model.accountmodel.UserModelResponse;
 import stories.model.shademodel.core.model.chatmodel.ChatErrorResponse;
 import stories.model.shademodel.core.model.chatmodel.ChatResponse;
-import stories.model.shademodel.core.model.jobmodel.JobFeedModelResponse;
-import stories.model.shademodel.core.model.jobmodel.JobModel;
-import stories.model.shademodel.core.model.jobmodel.JobModelResponse;
-import stories.model.shademodel.core.model.jobmodel.Location;
+import stories.model.shademodel.core.model.jobmodel.*;
 import stories.rest.APIFacade;
 import stories.rest.responsecheck.ResponseCheckFactory;
 import stories.test.BaseBackendTest;
@@ -36,6 +34,8 @@ public class HiredJobs extends BaseBackendTest {
     private APIFacade accountAPIFacade;
     public Integer userId2 = 0;
     public Integer userId3_1 = 0;
+    public Integer userId3_2 = 0;
+    public Integer userId3_3 = 0;
     public String request = "";
     public int jobId = 0;
 
@@ -56,7 +56,7 @@ public class HiredJobs extends BaseBackendTest {
     @AfterClass(dependsOnMethods = "shutdownBaseBackendTest")
     public void deleteUser()
             throws IOException, SQLException {
-/*        if (userId2 != 0) {
+       if (userId2 != 0) {
             accountAPIFacade.getAccountEndpoint().deleteUser(
                     userId2,
                     Collections.singletonList(ResponseCheckFactory.getStatusCodeCheck(200)),
@@ -82,8 +82,6 @@ public class HiredJobs extends BaseBackendTest {
                     "Deleting user with id: " + userId3_2 + ".");
         }
 
- */
-
     }
 
     @BeforeTest()
@@ -91,13 +89,13 @@ public class HiredJobs extends BaseBackendTest {
         settingsManager = SettingsManager.getSettingsManager();
         accountAPIFacade = new APIFacade(null, settingsManager.getDefaultBackendSettings());
         UserModelResponse response = null;
-/*
+
         String test_data = testDescription + "\n create user \n[ERROR] ";
         SoftAssert sa = new SoftAssert(test_data);
         //----Create Employer
         //Create the test user data without the DataProvider
         UserModel newUser2 = new UserModel();
-        newUser2.setEmail("test_applay_job_user2@distillery.com");
+        newUser2.setEmail("test_hired_job_user2@distillery.com");
         newUser2.setFirstName("First");
         newUser2.setLastName("Petriv");
         newUser2.setPassword("qqqaaa77");
@@ -114,7 +112,7 @@ public class HiredJobs extends BaseBackendTest {
         //  userId2 = 932;
         //----Create Employee
         UserModel newUser3 = new UserModel();
-        newUser3.setEmail("test_applay_job_user3_1@distillery.com");
+        newUser3.setEmail("test_hired_job_user3_1@distillery.com");
         newUser3.setFirstName("First");
         newUser3.setLastName("Petriv");
         newUser3.setPassword("qqqaaa77");
@@ -128,7 +126,7 @@ public class HiredJobs extends BaseBackendTest {
                 testDescription);
         userId3_1 = response.getAccess_token() != null ? response.getShadeUserModelResponse().getId() : 0;
 
-        newUser3.setEmail("test_applay_job_user3_2@distillery.com");
+        newUser3.setEmail("test_hired_job_user3_2@distillery.com");
         response = (UserModelResponse) accountAPIFacade.getAccountEndpoint().createUser(
                 newUser3,
                 Collections.singletonList(
@@ -136,24 +134,24 @@ public class HiredJobs extends BaseBackendTest {
                 testDescription);
         userId3_2 = response.getAccess_token() != null ? response.getShadeUserModelResponse().getId() : 0;
 
-        newUser3.setEmail("test_applay_job_user3_3@distillery.com");
+        newUser3.setEmail("test_hired_job_user3_3@distillery.com");
         response = (UserModelResponse) accountAPIFacade.getAccountEndpoint().createUser(
                 newUser3,
                 Collections.singletonList(
                         ResponseCheckFactory.getStatusCodeCheck(200)),
                 testDescription);
         userId3_3 = response.getAccess_token() != null ? response.getShadeUserModelResponse().getId() : 0;
-     */       userId2 = 6;
+     /*       userId2 = 6;
         userId3_1 = 7;
-
+*/
     }
 
 
 
-    @Test /*(description = "Applay Job",
-            dataProvider = "applayJob",
-             priority = 1))*/
-    public void applayJob() throws IOException, InterruptedException, SQLException {
+    @Test (/*description = "Applay Job",
+            dataProvider = "applayJob",*/
+             priority = 1)
+    public void hiredJob() throws IOException, InterruptedException, SQLException {
         accountAPIFacade = new APIFacade(null, settingsManager.getDefaultBackendSettings());
         settingsManager = SettingsManager.getSettingsManager();
 
@@ -174,7 +172,7 @@ public class HiredJobs extends BaseBackendTest {
         createJob.setEndDate("2017-07-30T09:06:53.932Z");
         createJob.setEndTime("2017-07-30T09:06:53.932Z");
         createJob.setSalary(100);
-        createJob.setSalaryType(1);
+        createJob.setSalaryType("1");
         createJob.setSummary("olololololo");
         createJob.setDescription("bububububu");
         createJob.setUserId(userId2);
@@ -208,33 +206,33 @@ public class HiredJobs extends BaseBackendTest {
         sa.assertNotNull(response_appled.get(0).getId());
 
 
-
         //  SoftAssert sa = new SoftAssert();
 
-        TimeUnit.SECONDS.sleep(600);
+     //   TimeUnit.SECONDS.sleep(600);
         //accempt employee wich applay the job
         request = "{\"selfId\":" + userId2 + ",\"employeeId\":" + userId3_1 + ",\"jobId\":" + responseJob.getId() + "}";
 
-        ChatResponse response_accept = (ChatResponse)accountAPIFacade.getEmployerEndpoint().acceptEmployee(
+        ChatResponse response_accept = (ChatResponse) accountAPIFacade.getEmployerEndpoint().acceptEmployee(
                 request,
                 Collections.singletonList(
                         ResponseCheckFactory.getStatusCodeCheck(200)),
                 testDescription);
+
+        sa = new SoftAssert();
         sa.assertNotNull(response_accept.getChat());
         sa.assertNotNull(response_accept.getChat().getId());
-        sa.assertNotNull(response_accept.getOpponentInfo().getId());
-        sa.assertNotNull(response_accept.getOpponentInfo().getFirstName());
-        sa.assertNotNull(response_accept.getJobInfo());
+ //       sa.assertNotNull(response_accept.getOpponentInfo().getId());
+ //       sa.assertNotNull(response_accept.getOpponentInfo().getFirstName());
+  //      sa.assertNotNull(response_accept.getJobInfo());
         sa.assertAll();
 
-        request = "{\"selfId\":" + userId3_1 + "\"jobId\":" + jobId + "}";
-}
-        Response response_acceptjob  = (Response) accountAPIFacade.getEmployeeEndpoint().acceptjob(
+        request = "{\"selfId\":" + userId3_1 + ",\"jobId\":" + jobId + "}";
+
+        Response response_acceptjob = (Response) accountAPIFacade.getEmployeeEndpoint().acceptjob(
                 request, userId3_1,
                 Collections.singletonList(
                         ResponseCheckFactory.getStatusCodeCheck(200)),
                 testDescription);
-
 
 
         //hiredJob
@@ -244,10 +242,43 @@ public class HiredJobs extends BaseBackendTest {
                 Collections.singletonList(
                         ResponseCheckFactory.getStatusCodeCheck(200)),
                 testDescription);
+        sa = new SoftAssert();
+        sa.assertTrue(response.get(0).getId() == jobId);
+        sa.assertNotNull(response.get(0).getName());
+        sa.assertTrue(response.get(0).getPoster().getId() == userId2 );
+//проверки
+        sa.assertAll();
+
 
 
     }
 
+    @Test (/*description = "Applay Job",
+            dataProvider = "applayJob",*/
+            priority = 2)
+    public void declinedJob() throws IOException, InterruptedException, SQLException {
+        accountAPIFacade = new APIFacade(null, settingsManager.getDefaultBackendSettings());
+        settingsManager = SettingsManager.getSettingsManager();
+
+        String request = "{\n" +
+                "  \"selfId\":" + userId3_1 +
+                ",  \"jobId\":" + jobId +
+                "}";
+        ChatErrorResponse response_decline = (ChatErrorResponse) accountAPIFacade.getEmployeeEndpoint().declinejob(
+                request,
+                Collections.singletonList(
+                        ResponseCheckFactory.getStatusCodeCheck(400)),
+                testDescription);
+
+        SoftAssert sa = new SoftAssert();
+        sa.assertTrue(response_decline.getErrorCode() == 308);
+        sa.assertTrue(response_decline.getErrorMessage().equals("The job has already been accepted by the employee."));
+//проверки
+        sa.assertAll();
+
+    }
 
 
 }
+
+
